Rearranging pipelines after the update
