+++ firebird-ng/src/app/data-pipelines/three-geometry.processor.ts
@@ -109,204 +109,9 @@ export function matchRulesToDetectors(ruleSets: DetectorThreeRuleSet[], detector
 
 export class ThreeGeometryProcessor {
 
-  rules: DetectorThreeRuleSet[] = [
-    {
-      names: ["FluxBarrel_env_25", "FluxEndcapP_26", "FluxEndcapN_28"],
-      rules: [
-        {
-          color: 0x373766,
-
-        }
-      ]
-    },
-    {
-      name: "EcalEndcapN*",
-      rules: [
-        {
-          patterns: ["**/crystal_vol_0"],
-          color: 0xffef8b,
-          material: new THREE.MeshStandardMaterial({
-            color: 0xffef8b,
-            roughness: 0.7,
-            metalness: 0.869,
-            transparent: true,
-            opacity: 0.8,
-            side: THREE.DoubleSide
-          })
-        },
-        {
-          patterns: ["**/inner_support*", "**/ring*"],
-          material: new THREE.MeshStandardMaterial({
-            color: 0x19a5f5,
-            roughness: 0.7,
-            metalness: 0.869,
-            transparent: true,
-            opacity: 0.8,
-            side: THREE.DoubleSide
-          })
-        }
-
-      ]
-    },
-    {
-      name: "InnerTrackerSupport_assembly_13",
-      rules: [
-        {
-          material: new THREE.MeshStandardMaterial({
-            color: 0xEEEEEE,
-            roughness: 0.7,
-            metalness: 0.3,
-            transparent: true,
-            opacity: 0.8,
-            blending: THREE.NormalBlending,
-            // premultipliedAlpha: true,
-            depthWrite: false, // Ensures correct blending
-            polygonOffset: true,
-            polygonOffsetFactor: 1,
-            side: THREE.DoubleSide
-          }),
-          outline: true,
-          outlineColor: 0x666666,
-          merge: true,
-          newName: "InnerTrackerSupport"
-        }
-      ]
-    },
-    {
-      name: "DIRC_14",
-      rules: [
-        {
-          patterns:     ["**/*box*", "**/*prism*"],
-          material: new THREE.MeshPhysicalMaterial({
-            color: 0xe5ba5d,
-            metalness: .9,
-            roughness: .05,
-            envMapIntensity: 0.9,
-            clearcoat: 1,
-            transparent: true,
-            //transmission: .60,
-            opacity: .6,
-            reflectivity: 0.2,
-            //refr: 0.985,
-            ior: 0.9,
-            side: THREE.DoubleSide,
-          }),
-          newName: "DIRC_barAndPrisms"
-        },
-        {
-          patterns: ["**/*rail*"],
-          newName: "DIRC_rails",
-          color: 0xAAAACC
-        },
-        {
-          patterns: ["**/*mcp*"],
-          newName: "DIRC_mcps"
-        }
-      ]
-
-    },
-    {
-      name: "VertexBarrelSubAssembly_3",
-      rules: [
-        {
-          merge: true,
-          outline: true
-        }
-      ]
-    },
-    {
-      name: "*",
-      rules: [
-        {
-          merge: true,
-          outline: true
-        }
-      ]
-    }
-  ]
-
-  calorimetry = new CalorimetryGeometryPrettifier();
-
-
-
-  glassMaterial = new THREE.LineBasicMaterial( {
-    color: 0xf1f1f1,
-    linewidth: 1,
-    linecap: 'round', //ignored by WebGLRenderer
-    linejoin:  'round' //ignored by WebGLRenderer
-  } );
-
-
-
-  params = {
-    alpha: 0.5,
-    alphaHash: true,
-    taa: true,
-    sampleLevel: 2,
-  };
-
-  vertexShader = `
-    varying vec2 vUv;
-    void main()	{
-      vUv = uv;
-      gl_Position = projectionMatrix * modelViewMatrix * vec4(position,1.0);
-    }
-  `;
-  fragmentShader = `
-		//#extension GL_OES_standard_derivatives : enable
-
-    varying vec2 vUv;
-    uniform float thickness;
-
-    float edgeFactor(vec2 p){
-    	vec2 grid = abs(fract(p - 0.5) - 0.5) / fwidth(p) / thickness;
-  		return min(grid.x, grid.y);
-    }
-
-    void main() {
-
-      float a = edgeFactor(vUv);
-
-      vec3 c = mix(vec3(1), vec3(0), a);
-
-      gl_FragColor = vec4(c, 1.0);
-    }
-  `;
-
-  shaderMaterial: THREE.ShaderMaterial;
-
-  alphaMaterial = new THREE.MeshStandardMaterial( {
-    color: 0xffffff,
-    alphaHash: this.params.alphaHash,
-    opacity: this.params.alpha
-  } );
-
-
-
-  // new MeshPhysicalMaterial({
-  //   color: 0xffff00, // Yellow color
-  //   metalness: 0,
-  //   roughness: 0,
-  //   transmission: 0.7, // High transparency
-  //   opacity: 1,
-  //   transparent: true,
-  //   reflectivity: 0.5
-  // });
 
   constructor() {
-    this.shaderMaterial = new THREE.ShaderMaterial({
-      uniforms: {
-        thickness: {
-          value: 1.5
-        }
-      },
-      vertexShader: this.vertexShader,
-      fragmentShader: this.fragmentShader
-    });
-  }
 
-  public process(detectors: Subdetector[]) {
-    this.processRuleSets(this.rules, detectors);
   }
 
   public processRuleSets(ruleSets: DetectorThreeRuleSet[], detectors: Subdetector[]) {
